+++ PDSProject/PDSProject/ServerDispatcher.cs
@@ -276,8 +276,8 @@ namespace PDSProject
                 INPUT input = JsonConvert.DeserializeObject<INPUT>(json);
                 float screenW = System.Windows.Forms.Screen.PrimaryScreen.Bounds.Width;
                 float screenH = System.Windows.Forms.Screen.PrimaryScreen.Bounds.Height;
-                input.mi.dx = (int)Math.Round(input.mi.dx * (65535 / screenW), 0);
-                input.mi.dy = (int)Math.Round(input.mi.dy * (65535 / screenH), 0);
+                input.mi.dx = (UInt16)Math.Round(input.mi.dx * (65535 / screenW), 0);
+                input.mi.dy = (UInt16)Math.Round(input.mi.dy * (65535 / screenH), 0);
 
                 server.Send(new byte[] { 0 }, socket);
                 INPUT[] inputList = { input };
